apply patch http://issues.apache.org/activemq/browse/AMQ-1293
git-svn-id: https://svn.apache.org/repos/asf/activemq/trunk@550476 13f79535-47bb-0310-9956-ffa450edef68

+++ assembly/src/test/java/org/apache/activemq/benchmark/BenchmarkSupport.java
@@ -184,7 +184,7 @@ public class BenchmarkSupport {
                 times++;
             }
             if (times > 0) {
-                average = total / times;
+                average = total / (double) times;
             }
 
             long oldtime = time;
